# swagger.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "NestJS기반 전자상거래 서비스"
    }
  ],
  "info": {
    "version": "1.0.0",
    "title": "E-Commerce Server"
  },
  "paths": {
...
    "/sign-in/google": {
      "post": {
        "tags": [
          "authentication"
        ],
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthUsecase.SignInBody"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUsecase.SignInResponse"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "description": "로그인 API\n\n구글 oauth2 인증을 통해 얻은 code를 body를 통해 제공해야 합니다.",
        "x-nestia-namespace": "sign-in.google.signInGoogle",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "authentication",
                "kind": "text"
              }
            ]
          }
        ]
      }
    },
    "/sign-in/github": {
      "post": {
        "tags": [
          "authentication"
        ],
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthUsecase.SignInBody"
              }
            }
          },
          "required": true,
          "x-nestia-encrypted": false
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUsecase.SignInResponse"
                }
              }
            },
            "x-nestia-encrypted": false
          }
        },
        "description": "로그인 API\n\n깃헙 oauth2 인증을 통해 얻은 code를 body를 통해 제공해야 합니다.",
        "x-nestia-namespace": "sign-in.github.signInGithub",
        "x-nestia-jsDocTags": [
          {
            "name": "tag",
            "text": [
              {
                "text": "authentication",
                "kind": "text"
              }
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthUsecase.SignInBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": false,
            "x-typia-required": true
          }
        },
        "nullable": false,
        "required": [
          "code"
        ],
        "x-typia-jsDocTags": []
      },
      "AuthUsecase.SignInResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "nullable": false,
            "x-typia-required": true
          },
          "refresh_token": {
            "type": "string",
            "nullable": false,
            "x-typia-required": true
          },
          "id_token": {
            "type": "string",
            "nullable": false,
            "x-typia-required": true
          }
        },
        "nullable": false,
        "required": [
          "access_token",
          "refresh_token",
          "id_token"
        ],
        "x-typia-jsDocTags": []
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}